requirements.txt
pip-chill==1.0.1
pysocks==1.7.1
python-dotenv==0.19.0
tweepy==3.10.0
